Add files via upload
Overview of the files I used in order to make a 3 tier web based application on AWS.

Amplify, Lambda, IAM, API Gateway and DynamoDB were the services I used to make this math application.  Amplify to deploy the application,, Lambda to make the JSON file in python and which dealt with the actual math aspect of the application, Dynamo DB would store the results after I'd call the Lambda function, used to IAM to align and assign roles to the services. Used RESTFUL Api to connect to my backend services which here was Lambda.